Widen revocation ticket to 64-bit; accept signed ints in json_require
The revocation-list ticket was a uint32, which can wrap or be rolled back by a disaster-recovery restore, silently making a client that is 'ahead' treat the list as unchanged and miss revocations. Widen it to int64_t in the C and C++ structs, docstrings, and JSON parse.

json_require previously accepted only unsigned integer types, so json_require<int64_t> tripped its trailing static_assert. Extend the integer branch to signed types by switching the type dispatch to STL concepts std::floating_point/std::integral/std::same_as (dropping the is_one_of type lists); the bool check moves ahead of the integral one since bool satisfies std::integral. Also drop the needless value-init of the local 'type'.

+++ include/session/pro_backend.h
@@ -187,7 +187,7 @@ typedef struct session_pro_backend_get_pro_revocations_request
         session_pro_backend_get_pro_revocations_request;
 struct session_pro_backend_get_pro_revocations_request {
     uint8_t version;
-    uint32_t ticket;
+    int64_t ticket;
 };
 
 typedef struct session_pro_backend_pro_revocation_item session_pro_backend_pro_revocation_item;
@@ -200,7 +200,7 @@ typedef struct session_pro_backend_get_pro_revocations_response
         session_pro_backend_get_pro_revocations_response;
 struct session_pro_backend_get_pro_revocations_response {
     session_pro_backend_response_header header;
-    uint32_t ticket;
+    int64_t ticket;
     /// Array of items, with items_count elements
     session_pro_backend_pro_revocation_item* items;
     size_t items_count;
